requirements.txt
asgiref==3.5.2
bcrypt==4.0.0
Django==4.1
django-browser-reload==1.6.0
django-ckeditor==6.5.0
django-ckeditor-5==0.1.8
django-dotenv==1.4.2
django-environ==0.9.0
django-guardian==2.4.0
django-js-asset==2.0.0
gunicorn==20.1.0
Jinja2==2.11.3
MarkupSafe==2.1.1
Pillow==9.2.0
psycopg2==2.9.3
python-dotenv==0.20.0
PyYAML==5.4.1
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
